Replace alpine images with scratch images
SPIRE is moving away from the alpine Docker release images in favor of
scratch images that contain only the release binary to minimize the size
of the images and include only the software that is necessary to run in
the container.

Also update SPIRE image versions in examples to the latest version since
the scratch images were only published as of v1.1.0.

Signed-off-by: Ryan Turner <turner@uber.com>